64 bit fix for caching sounds. Maybe nyXXXX will be happy now.
git-svn-id: https://svn.eduke32.com/eduke32@767 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/sounds.c
@@ -269,7 +269,7 @@ int loadsound(unsigned int num)
 
     g_sounds[num].lock = 200;
 
-    allocache((int *)&g_sounds[num].ptr,l,(char *)&g_sounds[num].lock);
+    allocache((intptr_t *)&g_sounds[num].ptr,l,(char *)&g_sounds[num].lock);
     kread(fp, g_sounds[num].ptr , l);
     kclose(fp);
     return 1;
